requirements.txt
# Dev dependencies
pytest==6.2.3
pytest-mock==3.5.1
requests-mock==1.9.2

# Dependencies
beautifulsoup4==4.8.2
CurrencyConverter==0.16
elasticsearch==7.8.0
fasttext==0.9.2
Flask==2.1.0
Flask-Bootstrap==3.3.7.1
lxml==4.6.3
multiprocess==0.70.12.2
numpy~=1.19.5
pandas==1.2.5
jsonlines==3.1.0
xlrd==2.0.1
openpyxl==3.0.7
pymongo==3.8.0
python-dateutil~=2.8.1
python-keystoneclient==4.0.0
python-swiftclient==3.9.0
redis==4.3.4
requests
retry~=0.9.2
rq==1.9.0
tokenizers==0.10.1
Unidecode==1.3.2
pysftp==0.2.9
pyzotero==1.5.5
timeout-decorator==0.5.0
boto3==1.26.41
sqlalchemy==1.4.31
psycopg2==2.9.3
python-dotenv==0.20.0
lmdb==1.2.1
cloudscraper==1.2.58
# Grobid client package
git+https://github.com/Barometre-de-la-Science-Ouverte/grobid_client_python.git#egg=grobid_client_python
# Softcite & Datastet client package
git+https://github.com/Barometre-de-la-Science-Ouverte/softdata_mentions_client.git#egg=softdata_mentions_client
